requirements.txt
fastapi
uvicorn
jinja2
python-multipart
azure-identity
azure-ai-vision-face
